Update release script to handle the case when one component has no updates.

+++ .misc/make-release.sh
@@ -4,6 +4,7 @@ set -x
 #Update all the README and checkin the results.
 
 #Define SCOWL_VERSION
+#Define PREV_VERSION
 
 test -e git/.git
 test -e git/scowl/README.in
@@ -17,6 +18,8 @@ test -e git/scowl/README.in
 
 #AGID:
   cd agid
+if [ -n "`git rev-list rel-$PREV_VERSION..HEAD .`" ]
+then
   git clean -xf .
   cd ..
   cp -a agid stage/agid-$SCOWL_VERSION
@@ -24,9 +27,15 @@ test -e git/scowl/README.in
   zip -q9rl agid-$SCOWL_VERSION.zip agid-$SCOWL_VERSION/
   tar cfz agid-$SCOWL_VERSION.tar.gz agid-$SCOWL_VERSION/
   cd ..
+else
+  cd ..
+  echo "SKIPPING AGID: No changes"
+fi
 
 #VarCon
   cd varcon
+if [ -n "`git rev-list rel-$PREV_VERSION..HEAD .`" ]
+then
   git clean -xf .
   cd ..
   cp -a varcon stage/varcon-$SCOWL_VERSION
@@ -34,9 +43,15 @@ test -e git/scowl/README.in
   zip -q9rl varcon-$SCOWL_VERSION.zip varcon-$SCOWL_VERSION/
   tar cfz varcon-$SCOWL_VERSION.tar.gz varcon-$SCOWL_VERSION/
   cd ..
+else
+  cd ..
+  echo "SKIPPING VarCon: No changes"
+fi
 
 #Alt12Dicts:
   cd alt12dicts
+if [ -n "`git rev-list rel-$PREV_VERSION..HEAD .`" ]
+then
   git clean -xf .
   cd ..
   cp -a alt12dicts stage/alt12dicts-$SCOWL_VERSION
@@ -44,6 +59,10 @@ test -e git/scowl/README.in
   zip -q9rl alt12dicts-$SCOWL_VERSION.zip alt12dicts-$SCOWL_VERSION/
   tar cfz alt12dicts-$SCOWL_VERSION.tar.gz alt12dicts-$SCOWL_VERSION/
   cd ..
+else
+  cd ..
+  echo "SKIPPING Alt12Dicts: No Changes"
+fi
 
 #SCOWL:
   cd scowl
