+++ .misc/push-release.sh
@@ -9,20 +9,35 @@ test -e release-bk
 rm -rf release
 cp -al release-bk release
 
-mkdir release/AGID/$SCOWL_VERSION
-cp git-fr/stage/agid-$SCOWL_VERSION.zip git-fr/stage/agid-$SCOWL_VERSION.tar.gz \
-   release/AGID/$SCOWL_VERSION
-cp git-fr/stage/agid-$SCOWL_VERSION/README release/AGID/$SCOWL_VERSION
+if [ -e git-fr/stage/agid-$SCOWL_VERSION.zip ]
+then
+  mkdir release/AGID/$SCOWL_VERSION
+  cp git-fr/stage/agid-$SCOWL_VERSION.zip git-fr/stage/agid-$SCOWL_VERSION.tar.gz \
+     release/AGID/$SCOWL_VERSION
+  cp git-fr/stage/agid-$SCOWL_VERSION/README release/AGID/$SCOWL_VERSION
+else
+  ( cd release/AGID; ln -s $PREV_VERSION $SCOWL_VERSION)
+fi
 
-mkdir release/VarCon/$SCOWL_VERSION
-cp git-fr/stage/varcon-$SCOWL_VERSION.zip git-fr/stage/varcon-$SCOWL_VERSION.tar.gz \
-   release/VarCon/$SCOWL_VERSION
-cp git-fr/stage/varcon-$SCOWL_VERSION/README release/VarCon/$SCOWL_VERSION
+if [ -e git-fr/stage/varcon-$SCOWL_VERSION.zip ]
+then
+  mkdir release/VarCon/$SCOWL_VERSION
+  cp git-fr/stage/varcon-$SCOWL_VERSION.zip git-fr/stage/varcon-$SCOWL_VERSION.tar.gz \
+     release/VarCon/$SCOWL_VERSION
+  cp git-fr/stage/varcon-$SCOWL_VERSION/README release/VarCon/$SCOWL_VERSION
+else
+  ( cd release/VarCon; ln -s $PREV_VERSION $SCOWL_VERSION)
+fi
 
-mkdir release/Alt12Dicts/$SCOWL_VERSION
-cp git-fr/stage/alt12dicts-$SCOWL_VERSION.zip git-fr/stage/alt12dicts-$SCOWL_VERSION.tar.gz \
-   release/Alt12Dicts/$SCOWL_VERSION
-cp git-fr/stage/alt12dicts-$SCOWL_VERSION/README release/Alt12Dicts/$SCOWL_VERSION
+if [ -e git-fr/stage/alt12dicts-$SCOWL_VERSION.zip ]
+then
+  mkdir release/Alt12Dicts/$SCOWL_VERSION
+  cp git-fr/stage/alt12dicts-$SCOWL_VERSION.zip git-fr/stage/alt12dicts-$SCOWL_VERSION.tar.gz \
+     release/Alt12Dicts/$SCOWL_VERSION
+  cp git-fr/stage/alt12dicts-$SCOWL_VERSION/README release/Alt12Dicts/$SCOWL_VERSION
+else
+  ( cd release/Alt12Dicts; ln -s $PREV_VERSION $SCOWL_VERSION)
+fi
 
 mkdir release/SCOWL/$SCOWL_VERSION
 cp git-fr/stage/scowl-$SCOWL_VERSION.zip git-fr/stage/scowl-$SCOWL_VERSION.tar.gz \
@@ -38,6 +53,6 @@ Please check everthing is in order and then
   1) ( cd release; rsync -v -n -a --delete . frs.sourceforge.net:/home/frs/project/wordlist/ )
   2) ( cd git-fr; git push -n git@github.com:kevina/wordlist.git master rel-$SCOWL_VERSION )
   3) ( cd git-fr/site; _tasks/deploy )
-  4) Go to https://sourceforge.net/projects/wordlist/files/SCOWL/2014.08.11.1/
+  4) Go to https://sourceforge.net/projects/wordlist/files/SCOWL/$SCOWL_VERSION/
      and update default version.
 EOF
